+++ app/src/main/java/com/survivalcoding/noteapp/presentation/addedit/AddEditViewModel.kt
@@ -2,7 +2,6 @@ package com.survivalcoding.noteapp.presentation.addedit
 
 import androidx.lifecycle.*
 import com.survivalcoding.noteapp.domain.model.Note
-import com.survivalcoding.noteapp.domain.repository.NoteRepository
 import com.survivalcoding.noteapp.domain.usecase.InsertNoteUseCase
 import kotlinx.coroutines.launch
 
@@ -12,16 +11,5 @@ class AddEditViewModel(private val insertNoteUseCase: InsertNoteUseCase) : ViewM
 
     fun insertNote(note: Note) = viewModelScope.launch {
         insertNoteUseCase(note)
-    }
-}
-
-@Suppress("UNCHECKED_CAST")
-class AddEditViewModelFactory(
-    private val repository: NoteRepository
-) : ViewModelProvider.NewInstanceFactory() {
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(AddEditViewModel::class.java))
-            return AddEditViewModel(InsertNoteUseCase(repository)) as T
-        else throw IllegalArgumentException()
     }
 }
